Added function getPairInfo

+++ contracts/SwapPair/interfaces/IRootSwapPairUpgradePairCode.sol
@@ -3,7 +3,7 @@ pragma solidity >= 0.6.0;
 
 interface IRootSwapPairUpgradePairCode {
     struct SwapPairCodeVersion {
-        uint8 contractCodeVersion;
+        uint contractCodeVersion;
     }
     
     function setSwapPairCode(TvmCell code, SwapPairCodeVersion codeVersion) external;
